requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.8.0
backcall==0.1.0
boto==2.49.0
boto3==1.9.158
botocore==1.12.158
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
docutils==0.14
Flask==1.0.3
gast==0.2.2
gensim==3.7.3
grpcio==1.21.1
h5py==2.9.0
idna==2.8
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
nltk==3.4.1
numpy==1.16.4
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
requests==2.22.0
s3transfer==0.2.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
smart-open==1.8.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==6.0.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
ConfigArgParse==0.14.0
tqdm==4.34.0
torch==1.2.0
tensorboardX==1.8
